Polish guides and tools UX, align full report with example layout.
Guide index and article pages get richer cards, callouts, and analysis CTAs. Calculator tools get fixed form overflow, before/after result panels, and tighter spacing. The unlocked analysis report now shares the example report grid and sections. Floor validation accepts decimal våning values.

+++ lib/schemas.ts
@@ -12,7 +12,7 @@ export const CreateAnalysisSchema = z.object({
 
   livingAreaSqm: z.coerce.number().positive().optional(),
   rooms: z.coerce.number().positive().optional(),
-  floor: z.coerce.number().int().optional(),
+  floor: z.coerce.number().optional(),
   totalFloors: z.coerce.number().int().optional(),
   hasBalcony: z.coerce.boolean().optional().default(false),
   balconyDirection: z.string().max(50).optional(),
